+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# Version 1.5.0
+
+## Bugfixes
+
+* None
+
+## Features
+
+* Extend session bean descriptor with remove method information
+* Add new @Remote annotation class implementation
+
 # Version 1.4.0
 
 ## Bugfixes